Move search StateCheckBox to containers
